refactor(lib): extract lib/ip.php (ADR-004 foundation)
19 functions moved byte-identically from lib.php to lib/ip.php:
ipam_bind_binary, ip_in_any_cidr, parse_cidr, apply_prefix_mask,
ipam_compute_broadcast_bin, ipam_compute_gateway_bin, ip_in_cidr,
normalize_ip, ipv4_bin_to_int, ipv4_int_to_bin, ipv4_int_to_text,
ipv4_assignable_count, subnet_contains_bin, ipv4_broadcast_bin,
ipv4_broadcast_int, ipv6_bin_increment, netmask_to_prefix,
cidr_from_ip_and_prefix, subnet_overlap_warning_text.

Pattern mirrors Task 2.1 (lib/utils.php at 57710217): pure file move,
no callsite changes (functions stay in global namespace per ADR-004
Option E), dual require_once in init.php + lib.php for direct-bootstrap
callers (tests/bootstrap.php, api.php, status.php), parity test via
ReflectionFunction.

ipam_bind_binary moved byte-identically; invariant #1 (binary IPs at
native 4B/16B + PARAM_LOB) preserved. Manual ip-binary-auditor pass:

- Rule 1 (native length, never padded): clean — pack('N',...) for IPv4
  and pack('C16',...) for IPv6; no left-pad, no str_repeat("\\0").
- Rule 2 (ipam_bind_binary for binary binds): function body
  byte-identical to pre-move (diff -> BYTE_IDENTICAL).
- Rule 3 (normalize_ip before shell/socket): no new sinks added.
- Rule 4 (reserved-IP skip): no scan paths touched.
- Rule 5 (round-trip test vectors): BinaryBindTest passes in full
  phpunit run.
- Rule 6 (no ip_version on addresses): no INSERT/UPDATE to addresses.

Deferred to v3.32.0 (lib/subnets.php / lib/addresses.php): the DB-
dependent helpers in the same neighbourhood — build_subnet_tree,
subnet_direct_counts, subnet_aggregated_counts, ipv4_unassigned_summary,
ipv4_unassigned_aggregated, find_next_available_ipv4,
ipv6_enumerate_first_n, find_containing_subnet, ensure_subnet_exists,
detect_subnet_overlaps. normalize_status stays in lib.php (string
status normaliser, not IP math).

Local gate: php -l (3/3 pass), phpstan L9 (0 errors), phpcs (clean),
phpunit (1233/1233 pass), lib-module-linter (exit 0).

Sub-task of #907. Refs #1222.
